requirements.txt
matplotlib==3.5.1
numpy==1.23.5
Pillow==10.0.0
scikit_learn==1.3.0
torch==1.10.0
torchvision==0.11.0
